requirements.txt
aiogram
python-dotenv
psycopg2-binary
boto3
pdf2image
Pillow
